feat: agregar endpoint para confirmar una recepcion

+++ src/main/java/com/mrdevv/service/IRecepcionService.java
@@ -8,7 +8,7 @@ public interface IRecepcionService {
 
     ResponseRecepcionDTO saveRecepcion(CreateRecepcionDTO recepcionDTO);
 
-    Recepcion findRecepcionById(Integer recepcionId);
+    Recepcion findRecepcionById(Long recepcionId);
 
     void confirmarRecepcion(Long recepcionId);
